agents/04_code_interpreter_bot/requirements.txt
# Core Bedrock AgentCore runtime (AWS-managed agent hosting)
bedrock-agentcore
strands-agents
